# Manually implement SSZ for Checkpoint