--- a/consensus/types/src/checkpoint.rs
+++ b/consensus/types/src/checkpoint.rs
@@ -1,7 +1,7 @@
 use crate::test_utils::TestRandom;
 use crate::{Epoch, Hash256};
 use serde::{Deserialize, Serialize};
-use ssz_derive::{Decode, Encode};
+use ssz::{Decode, DecodeError, Encode};
 use test_random_derive::TestRandom;
 use tree_hash_derive::TreeHash;
 
@@ -19,14 +19,70 @@ use tree_hash_derive::TreeHash;
     Hash,
     Serialize,
     Deserialize,
-    Encode,
-    Decode,
     TreeHash,
     TestRandom,
 )]
 pub struct Checkpoint {
     pub epoch: Epoch,
     pub root: Hash256,
+}
+
+/// Use a custom implementation of SSZ to avoid the overhead of the derive macro.
+impl Encode for Checkpoint {
+    fn is_ssz_fixed_len() -> bool {
+        true
+    }
+
+    fn ssz_append(&self, buf: &mut Vec<u8>) {
+        self.epoch.ssz_append(buf);
+        self.root.ssz_append(buf);
+    }
+
+    #[allow(clippy::arithmetic_side_effects)]
+    fn ssz_bytes_len(&self) -> usize {
+        self.epoch.ssz_bytes_len() + self.root.ssz_bytes_len()
+    }
+}
+
+/// Use a custom implementation of SSZ to avoid the overhead of the derive macro.
+impl Decode for Checkpoint {
+    fn is_ssz_fixed_len() -> bool {
+        true
+    }
+
+    #[allow(clippy::arithmetic_side_effects)]
+    fn ssz_fixed_len() -> usize {
+        <Epoch as Decode>::ssz_fixed_len() + <Hash256 as Decode>::ssz_fixed_len()
+    }
+
+    fn from_ssz_bytes(bytes: &[u8]) -> Result<Self, ssz::DecodeError> {
+        let expected = <Self as Decode>::ssz_fixed_len();
+
+        let (epoch, root) = bytes
+            .split_at_checked(<Epoch as Decode>::ssz_fixed_len())
+            .ok_or(DecodeError::InvalidByteLength {
+                len: bytes.len(),
+                expected,
+            })?;
+
+        if root.len() != <Hash256 as Decode>::ssz_fixed_len() {
+            return Err(DecodeError::InvalidByteLength {
+                len: bytes.len(),
+                expected,
+            });
+        }
+
+        let epoch = {
+            let mut array = [0; 8];
+            array.copy_from_slice(epoch);
+            u64::from_le_bytes(array)
+        };
+
+        Ok(Self {
+            epoch: Epoch::new(epoch),
+            root: Hash256::from_slice(root),
+        })
+    }
 }
 
 #[cfg(test)]