@paulhauner 2024-12-05
Here's the SSZ decode implementation, for reference:

```rust
fn from_ssz_bytes(bytes: &[u8]) -> std::result::Result<Self, ssz::DecodeError> {
    if <Self as ssz::Decode>::is_ssz_fixed_len() {
        if bytes.len() != <Self as ssz::Decode>::ssz_fixed_len() {
            return Err(ssz::DecodeError::InvalidByteLength {
                len: bytes.len(),
                expected: <Self as ssz::Decode>::ssz_fixed_len(),
            });
        }
        let mut start: usize = 0;
        let mut end = start;
        let epoch = {
            start = end;
            end = end
                .checked_add(<Epoch as ssz::Decode>::ssz_fixed_len())
                .ok_or_else(|| ssz::DecodeError::OutOfBoundsByte {
                    i: usize::max_value(),
                })?;
            let slice = bytes
                .get(start..end)
                .ok_or_else(|| ssz::DecodeError::InvalidByteLength {
                    len: bytes.len(),
                    expected: end,
                })?;
            <Epoch as ssz::Decode>::from_ssz_bytes(slice)?
        };
        let root = {
            start = end;
            end = end
                .checked_add(<Hash256 as ssz::Decode>::ssz_fixed_len())
                .ok_or_else(|| ssz::DecodeError::OutOfBoundsByte {
                    i: usize::max_value(),
                })?;
            let slice = bytes
                .get(start..end)
                .ok_or_else(|| ssz::DecodeError::InvalidByteLength {
                    len: bytes.len(),
                    expected: end,
                })?;
            <Hash256 as ssz::Decode>::from_ssz_bytes(slice)?
        };
        Ok(Self { epoch, root })
    } else {
        let mut builder = ssz::SszDecoderBuilder::new(bytes);
        builder.register_type::<Epoch>()?;
        builder.register_type::<Hash256>()?;
        let mut decoder = builder.build()?;
        let epoch = decoder.decode_next()?;
        let root = decoder.decode_next()?;
        Ok(Self { epoch, root })
    }
}
```