@paulhauner 2024-12-05
Closing this in favour of implementing the optimisation at the derive macro level: https://github.com/sigp/ethereum_ssz/pull/40